requirements.txt
bleach==1.4.3
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
django-configstore==0.3
django-el-pagination==2.1.2
django-html-sanitizer==0.1.5
django-mptt==0.8.4
django-registration==2.1
docopt==0.4.0
Embedly==0.5.0
gunicorn==19.6.0
html5lib==0.9999999
httplib2==0.9.2
mailchimp==2.0.9
pycrypto==2.6
python-instagram==1.3.2
pytz==2016.4
requests==2.10.0
simplejson==3.8.2
six==1.10.0
static3==0.7.0
